슬라이드 수정
이제 터치와 꾹 누르기를 분리하고, 공중 꾹 누르기는 높이를 고정하게 할 예정. 판정을 맡는 콜라이더는 플레이어 캐릭터 하위에 넣을 예정.

+++ 겨울프로젝트/Assets/Characters/Scripts/PlayerManager.cs
@@ -7,7 +7,7 @@ public class PlayerManager : MonoBehaviour
     public static PlayerManager instance;
 
     private GameObject player;
-    public float BPM = 100;
+    public int BPM = 100;
     public float floorSpeed = 1;
 
     public int score = 0;
